# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen",
        "description": "This is Documentation of Lumen",
        "contact": {
            "email": "brayenprayoga41@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/user/master-menu": {
            "get": {
                "tags": [
                    "Master Menu"
                ],
                "operationId": "45cdb07d74111abaa4ad6e413eb506b3",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-menu/simpan": {
            "post": {
                "tags": [
                    "Master Menu"
                ],
                "operationId": "b53ca90e67e337ddfb50baa8720231f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_menu": {
                                        "title": "Kode Menu",
                                        "description": "Kode Menu",
                                        "example": ""
                                    },
                                    "nama": {
                                        "title": "nama",
                                        "description": "nama",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-menu/edit": {
            "post": {
                "tags": [
                    "Master Menu"
                ],
                "operationId": "d3389775c0e361f59c81d49001859cef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_menu": {
                                        "title": "Kode Menu",
                                        "description": "Kode Menu",
                                        "example": ""
                                    },
                                    "nama": {
                                        "title": "nama",
                                        "description": "nama",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-menu/hapus": {
            "get": {
                "tags": [
                    "Master Menu"
                ],
                "operationId": "ee44c2ec581e917faa851f9967323063",
                "parameters": [
                    {
                        "name": "kode_menu",
                        "in": "query",
                        "description": "Kode menu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-ukuran-layar": {
            "get": {
                "tags": [
                    "Master Ukuran Layar"
                ],
                "operationId": "9055212e63e533ed3e280160c65caca3",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-ukuran-layar/simpan": {
            "post": {
                "tags": [
                    "Master Ukuran Layar"
                ],
                "operationId": "1f25d55f745e1dc77900417ccc8ecb02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "width": {
                                        "title": "Width",
                                        "description": "Width",
                                        "example": ""
                                    },
                                    "height": {
                                        "title": "Height",
                                        "description": "Height",
                                        "example": ""
                                    },
                                    "kode_dealer": {
                                        "title": "kode_dealer",
                                        "description": "kode_dealer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-ukuran-layar/edit": {
            "post": {
                "tags": [
                    "Master Ukuran Layar"
                ],
                "operationId": "31b26363a2b379b10c950d5627145110",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "width": {
                                        "title": "Width",
                                        "description": "Width",
                                        "example": ""
                                    },
                                    "height": {
                                        "title": "Height",
                                        "description": "Height",
                                        "example": ""
                                    },
                                    "kode_dealer": {
                                        "title": "kode_dealer",
                                        "description": "kode_dealer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-ukuran-layar/hapus": {
            "get": {
                "tags": [
                    "Master Ukuran Layar"
                ],
                "operationId": "2cb63e0632260fad6c48d2c846d7d275",
                "parameters": [
                    {
                        "name": "kode_dealer",
                        "in": "query",
                        "description": "Kode Dealer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Hubungi Developer Untuk Akun Login",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    }
}